pyproject.toml
[project]
name = "squackit"
version = "0.2.0"
description = "Semi-QUalified Agent Companion Kit — the stateful intelligence + MCP server layer for fledgling-equipped agents."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [
    { name = "Teague Sterling" },
]
keywords = ["mcp", "ai-agents", "code-intelligence", "fledgling", "duckdb"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "ast-pluckit>=0.7.0",
    "fastmcp>=3.0",
]

[project.scripts]
squackit = "squackit.server:main"

[project.optional-dependencies]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocs-exclude>=1.0",
    "pymdown-extensions>=10.9",
]

[project.urls]
Homepage = "https://github.com/teaguesterling/squackit"
Documentation = "https://squackit.readthedocs.io"
Repository = "https://github.com/teaguesterling/squackit"
Issues = "https://github.com/teaguesterling/squackit/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["squackit"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
test = [
    "pytest>=7.0",
]
